requirements.txt
Brotli==1.0.9
click==7.1.2
cycler==0.10.0
dash==1.19.0
dash-bootstrap-components==0.11.3
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mysql-connector==2.2.9
nltk==3.5
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.1
plotly==4.14.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
retrying==1.3.3
six==1.15.0
tqdm==4.58.0
Werkzeug==1.0.1
wordcloud==1.8.1
